created logic in case of no rooms and fixed a bug which led to rooms not being...

See merge request m6870/a5399/trivia-project!57

+++ Trivia Project/Backend/include/responses/GetGameResultsResponse.h
@@ -8,7 +8,7 @@ struct PlayerResults
 	std::string username;
 	unsigned int correctAnswerCount;
 	unsigned int wrongAnswerCount;
-	unsigned int averageAnswerTime;
+	float averageAnswerTime;
 
 	NLOHMANN_DEFINE_TYPE_INTRUSIVE(PlayerResults, username, correctAnswerCount, wrongAnswerCount, averageAnswerTime);
 };
